updated some files

+++ src/deliver.c
@@ -172,7 +172,7 @@ void lamb_event_loop(void) {
 void *lamb_push_loop(void *arg) {
     int err;
     int fd, rc;
-    long long timeout;
+    int timeout;
     lamb_req_t *client;
     lamb_queue_t *queue;
     
@@ -277,7 +277,7 @@ void *lamb_push_loop(void *arg) {
 void *lamb_pull_loop(void *arg) {
     int err;
     int fd, rc;
-    long long timeout;
+    int timeout;
     lamb_node_t *node;
     lamb_req_t *client;
     lamb_queue_t *queue;
